fix(ui): cleanup route summary display
