+++ renderer/components/Activity/PaymentModal/index.js
@@ -1,3 +1,4 @@
 export PaymentModal from './PaymentModal'
 export Route from './Route'
 export Htlc from './Htlc'
+export HtlcHops from './HtlcHops'
